Use a common protein-checmical type to simplify the user selection of interaction type
Ref : Tweaks to the interaction wizard #275
